Split pop common and service code
