+++ build_msvc/vbitcoin-tx/vbitcoin-tx.vcxproj
@@ -27,20 +27,11 @@
     <ProjectReference Include="..\libbitcoin_crypto\libbitcoin_crypto.vcxproj">
       <Project>{6190199c-6cf4-4dad-bfbd-93fa72a760c1}</Project>
     </ProjectReference>
-    <ProjectReference Include="..\libbitcoin_server\libbitcoin_server.vcxproj">
-      <Project>{460fee33-1fe1-483f-b3bf-931ff8e969a5}</Project>
-    </ProjectReference>
-    <ProjectReference Include="..\libleveldb\libleveldb.vcxproj">
-      <Project>{18430fef-6b61-4c53-b396-718e02850f1b}</Project>
-    </ProjectReference>
     <ProjectReference Include="..\libbitcoin_util\libbitcoin_util.vcxproj">
       <Project>{b53a5535-ee9d-4c6f-9a26-f79ee3bc3754}</Project>
     </ProjectReference>
     <ProjectReference Include="..\libunivalue\libunivalue.vcxproj">
       <Project>{5724ba7d-a09a-4ba8-800b-c4c1561b3d69}</Project>
-    </ProjectReference>
-    <ProjectReference Include="..\libleveldb\libleveldb.vcxproj">
-      <Project>{18430fef-6b61-4c53-b396-718e02850f1b}</Project>
     </ProjectReference>
     <ProjectReference Include="..\libsecp256k1\libsecp256k1.vcxproj">
       <Project>{bb493552-3b8c-4a8c-bf69-a6e7a51d2ea6}</Project>
