+++ src/chainparams.cpp
@@ -16,7 +16,7 @@
 
 #include <assert.h>
 
-#include <vbk/genesis.hpp>
+#include <vbk/genesis_common.hpp>
 
 #include <boost/algorithm/string/classification.hpp>
 #include <boost/algorithm/string/split.hpp>
